backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
pydantic==2.8.0
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
yfinance==0.2.43
requests==2.32.0
pandas==2.2.2
python-multipart==0.0.9
websockets==12.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7